12.15更新 版本2.6.2 1. 修复了直接输入Ultipay的报错 2. 增加了小数位的支持！

+++ src/main/java/com/minecraft/economy/apis/UltiEconomyAPI.java
@@ -39,7 +39,7 @@ public interface UltiEconomyAPI {
      * @param player_name 玩家名
      * @return 玩家的现金数量 ，如果没有找到玩家就返回-1
      */
-    public Integer checkMoney(String player_name);
+    public Double checkMoney(String player_name);
 
     /**
      * 获取玩家存款
@@ -47,7 +47,7 @@ public interface UltiEconomyAPI {
      * @param player_name 玩家名
      * @return 玩家的存款数量 ，如果没有找到玩家就返回-1
      */
-    public Integer checkBank(String player_name);
+    public Double checkBank(String player_name);
 
     /**
      * 添加金额到玩家现金
@@ -56,7 +56,7 @@ public interface UltiEconomyAPI {
      * @param amount      数额，请不要输入负数
      * @return 是否添加成功 boolean
      */
-    public Boolean addTo(String player_name, Integer amount);
+    public Boolean addTo(String player_name, Double amount);
 
     /**
      * 添加金额到玩家存款
@@ -65,7 +65,7 @@ public interface UltiEconomyAPI {
      * @param amount      数额，请不要输入负数
      * @return 是否添加成功 boolean
      */
-    public Boolean addToBank(String player_name, Integer amount);
+    public Boolean addToBank(String player_name, Double amount);
 
     /**
      * 从玩家现金中取钱
@@ -74,7 +74,7 @@ public interface UltiEconomyAPI {
      * @param amount      数额，请不要输入负数
      * @return 是否取走成功 boolean
      */
-    public Boolean takeFrom(String player_name, Integer amount);
+    public Boolean takeFrom(String player_name, Double amount);
 
     /**
      * 从玩家银行中取钱
@@ -83,7 +83,7 @@ public interface UltiEconomyAPI {
      * @param amount      数额，请不要输入负数
      * @return 是否取走成功 boolean
      */
-    public Boolean takeFromBank(String player_name, Integer amount);
+    public Boolean takeFromBank(String player_name, Double amount);
 
     /**
      * 玩家间转账
@@ -93,7 +93,7 @@ public interface UltiEconomyAPI {
      * @param amount 数额
      * @return 是否转账成功 boolean
      */
-    public Boolean transferMoney(String payer, String payee, Integer amount);
+    public Boolean transferMoney(String payer, String payee, Double amount);
 
     /**
      * 从一个玩家的现金转钱到存款
@@ -102,7 +102,7 @@ public interface UltiEconomyAPI {
      * @param amount      数额
      * @return 是否成功 boolean
      */
-    public Boolean transferMoneyToBank(String player_name, Integer amount);
+    public Boolean transferMoneyToBank(String player_name, Double amount);
 
     /**
      * 从一个玩家的存款转钱到现金
@@ -111,5 +111,5 @@ public interface UltiEconomyAPI {
      * @param amount      数额
      * @return 是否成功 boolean
      */
-    public Boolean transferBankToMoney(String player_name, Integer amount);
+    public Boolean transferBankToMoney(String player_name, Double amount);
 }
